lab5/requirements.txt
fastapi
uvicorn
sqlalchemy
redis
python-jose[cryptography]
passlib[bcrypt]
psycopg2-binary
python-multipart
